web/requirements.txt
flask
Flask-PyMongo
uwsgi

